services/private_github_crawler/requirements.txt
gitpython
stdlib-list
pip
PyGithub
python-magic
redis
rq
rsyslog
